Fixed my error where I was using a local session variable rather than the proper user info for the weapon bob speed.

+++ src/d_player.h
@@ -337,7 +337,7 @@ struct userinfo_t : TMap<FName,FBaseCVar *>
 	{
 		return *static_cast<FFloatCVar *>(*CheckKey(NAME_StillBob));
 	}
-	double GetWBobSpeed() const
+	float GetWBobSpeed() const
 	{
 		return *static_cast<FFloatCVar *>(*CheckKey(NAME_WBobSpeed));
 	}
